+++ MessageFlow/20-config.mqsc
@@ -2,3 +2,4 @@
 DEFINE QLOCAL('ORDER.REQUEST') REPLACE
 DEFINE QLOCAL('ORDER.RESPONSE') REPLACE
 DEFINE QLOCAL('ORDER.RESTAURANT') REPLACE
+DEFINE QLOCAL('ORDER.TEST') REPLACE
