+++ MessageFlow/src/main/java/com/ibmmq/messageflow/studyTests/test02/MessageFlowApplication.java
@@ -1,7 +1,6 @@
-package com.ibmmq.messageflow.test02;
+package com.ibmmq.messageflow.studyTests.test02;
 
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
