+++ MessageFlow/src/main/java/com/ibmmq/messageflow/studyTests/test04/OrderProducer.java
@@ -1,11 +1,17 @@
-package com.ibmmq.messageflow.test03;
+package com.ibmmq.messageflow.studyTests.test04;
 
 import jakarta.jms.*;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jms.annotation.JmsListener;
+import org.springframework.jms.core.JmsTemplate;
+import org.springframework.jms.core.MessageCreator;
 import org.springframework.stereotype.Component;
 
 @Component
 public class OrderProducer {
+
+    @Autowired
+    private JmsTemplate jmsTemplate;
 
     @JmsListener(destination = Requester.ORDER_QUEUE)
     public void onMessage(Message msg, Session session){
@@ -32,9 +38,21 @@ public class OrderProducer {
             TextMessage replyMsg = session.createTextMessage("Replying to " + text);
             replyMsg.setJMSCorrelationID(msgID);
             replyDest.send(replyMsg);
+            sendMessageToAnotherQueue(((TextMessage) msg).getText());
         } catch (JMSException e) {
             e.printStackTrace();
         }
 
+    }
+
+
+    private void sendMessageToAnotherQueue(String payload) {
+        jmsTemplate.send("DEV.QUEUE.1", new MessageCreator() {
+            @Override
+            public Message createMessage(Session session) throws JMSException {
+                TextMessage message = session.createTextMessage(payload);
+                return message;
+            }
+        });
     }
 }
