+++ MessageFlow/src/main/resources/application.properties
@@ -9,6 +9,5 @@ ibm.mq.password=passw0rd
 #ibm.mq.receivingQueueName=DEV.TO.Q1
 
 #queues
-queue.name01=ORDER.REQUEST
-queue.name02=ORDER.RESPONSE
 order.queue=ORDER.RESTAURANT
+test.queue=ORDER.TEST
